Introduce `ValidatedUrlData` that is returned by the `ValidatedUrlDataProvider`
